==> all_code/1 Add_century final.ipynb <==
import re
import pandas as pd
import os

df = pd.read_csv('totalset.csv', compression = 'gzip')
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

# df_1 = : 34807
# df_2 = 34808 tm 36374 
# df_3 = 36375 tm 40374
# df_4 = 40375 tm 40593
# df_5 = 40594 :

df_1 = df[:34808]
df_2 = df[34808:36375] 
df_3 = df[36375:40375]
df_4 = df[40375:40594]
df_5 = df[40594:]

# for dataset 1: everything is 1600s
meta_data2 = pd.read_excel('Metadata_IMPACT_books.xlsx')
meta_data3 = pd.read_excel('MetadataHistorical_newspaper_OCR_groundtruth.xlsx') 
meta_data4 = pd.read_excel('Metadata_DBNL_OCR_v1.xlsx')
# for dataset 5: everything is 1900s

print(meta_data2.columns)
print(meta_data3.columns)
print(meta_data4.columns)

print('meta_data2:', str(meta_data2['Did'].head(1))) # Impact Books
print('meta_data3:', str(meta_data3['Identifier'].head(1))) # newspaper pages
print('meta_data4:', str(meta_data4['ti_id'].head(1))) # DBNL books

#datasets with constant values for year
df_1_year = df_1
df_5_year = df_5
df_1_year['year'] = '16--'
df_5_year['year'] = '1937'

print(meta_data2.head(1))

meta_data2['year'] = meta_data2['Dyear']
meta_data2['identifier'] = meta_data2['Did'].astype(str)
meta_data2 = meta_data2[['identifier', 'year']]

df_2_year = pd.merge(left=df_2, right=meta_data2, on='identifier')
df_2_year.head(1)

print(meta_data3.head(1))

print(meta_data3.columns)

df_3.head()

# df_3: DDD_010197037_001.jp2_ocr
# df_3: MMGARO01_000177181_002_master.jp2_ocr
# metadata3: ddd:010860580:mpeg21
# metadata3: MMKB08:000192440:mpeg21

df_3= df_3[~df_3.identifier.str.contains("_reocr")].reset_index(drop = True)

len(df_3)

def shorten(identifier):
    uniform = identifier.replace('_', ':')
    splitted = uniform.split(':')
    new_id = (':'.join(splitted[:2])).lower()
    return new_id 

df_3['short id'] = df_3['identifier'].apply(lambda x: shorten(x))
meta_data3['short id'] = meta_data3['Identifier'].apply(lambda x: shorten(x))


# df_3: DDD_010197037_001.jp2_ocr
# df_3: MMGARO01_000177181_002_master.jp2_ocr
# metadata3: ddd:010860580:mpeg21
# metadata3: MMKB08:000192440:mpeg21

meta_data3.head()

df_3.head()

meta_data3['year'] = meta_data3['Date of publication']
meta_data3['identifier'] = meta_data3['Identifier']
meta_data3 = meta_data3[['short id', 'year']]

df_3_year = pd.merge(left=df_3, right=meta_data3, on='short id')
df_3_year = df_3_year.drop("short id", axis=1)
df_3_year.head(1)


len(df_3_year)

df_4.head(1)

meta_data4.head(1)

meta_data4['year'] = meta_data4['jaar']
meta_data4['identifier'] = meta_data4['ti_id'].astype(str)
meta_data4 = meta_data4[['identifier', 'year']]

df_4_year = pd.merge(left=df_4, right=meta_data4, on='identifier')
df_4_year.head(1)

df_1_year.head(1)

df_year = pd.concat([df_1_year, df_2_year, df_3_year, df_4_year, df_5_year])

#length should be len(df)-2000
print(len(df) - 2000)
print(len(df_year))

df_year['century'] = df_year['year'].apply(lambda x: str(x)[:2] + '00s')

print(df_year.head())

print(df_year['century'].unique())

#df_17th = df_year[df_year['century'] == '1600s']
#df_18th = df_year[df_year['century'] == '1700s']
#df_19th = df_year[df_year['century'] == '1800s']
#df_20th = df_year[df_year['century'] == '1900s']

#df_17th.head

df_year.to_csv('df_centuries.csv', index=True)
